refactor : 리액션 취소 로직 리팩토링

+++ core/domain/src/main/java/com/barlow/core/domain/reaction/ReactionProcessor.java
@@ -23,6 +23,9 @@ public class ReactionProcessor {
 
 	@Transactional
 	public void removeReaction(User user, Reaction reaction) {
+		if (reactionRepository.retrieve(user, reaction) == null) {
+			throw ReactionDomainException.alreadyRemoved(reaction.getTargetType().name(), reaction.getTargetId());
+		}
 		reactionRepository.removeReaction(user, reaction);
 	}
 }
